added synthetic dataset with level settings: level 1: 1 feature; level 2: 3 features; level 3: 1 feature & 10% randome label flipping; level 4: 3 features & 25% random regression model weight dropout; level 5: 3 features & 10% random label flipping & 25% weight dropout. A total of 5 feature types exist, level 1 and 3 has 4 distracting feature types, level 2, 4, 5 has 2 distracting feature types

+++ synthetic_dataset_generation/synthetic_dataset_generation.py
@@ -305,8 +305,12 @@ class SyntheticDatasetGenerator:
         output_dir = os.path.dirname(self.output_path)
         if not os.path.exists(output_dir):
             os.makedirs(output_dir)
-        with open(os.path.join(output_dir, "clf_model.pkl"), "wb") as f:
-            pickle.dump(self.clf_model, f)
+        if not os.path.exists(os.path.join(output_dir, "clf_model.pkl")):
+            with open(os.path.join(output_dir, "clf_model.pkl"), "wb") as f:
+                pickle.dump(self.clf_model, f)
+        else: 
+            with open(os.path.join(output_dir, "clf_model.pkl"), "rb") as f:
+                self.clf_model = pickle.load(f)
         return self.clf_model
 
     def initialize_from_metadata(self, metadata_path: str):
